+++ CHANGELOG.md
@@ -54,7 +54,9 @@
 
 ## [Stable 7.2.7] - 2026-01-08
 ### Fixed
+- History sync should work properly now (was only working for freshmen)
 - Remove error UI when a class is set to point-based
+- Font of 'Graderoom' text is no longer less bold on main page than every other page
 
 ## [Stable 7.2.6] - 2026-01-07 - Happy New Year!
 ### Added